# api/src/main/java/com/app/oopsly/api/viewmodel/TestSuiteReq.java
/*
 *    Copyright 2026 Hao Nguyen Tan
 *
 *    Licensed under the Apache License, Version 2.0 (the "License");
 *    you may not use this file except in compliance with the License.
 *    You may obtain a copy of the License at
 *
 *        http://www.apache.org/licenses/LICENSE-2.0
 *
 *    Unless required by applicable law or agreed to in writing, software
 *    distributed under the License is distributed on an "AS IS" BASIS,
 *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *    See the License for the specific language governing permissions and
 *    limitations under the License.
 */

package com.app.oopsly.api.viewmodel;

import com.app.oopsly.api.entity.TestSuiteSelectionPayload;
import jakarta.validation.constraints.NotNull;
import java.util.List;
import java.util.UUID;
import lombok.NonNull;

public record TestSuiteReq(
        @NonNull @NotNull String title,
        Boolean isActive,
        List<UUID> subjectIds,
        TestSuiteSelectionPayload selection) {}