# api/src/main/resources/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:9009/api/v1/oopsly",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/test-suites/{testSuiteId}/questions/{id}": {
      "get": {
        "tags": [
          "Question"
        ],
        "summary": "Get question by ID",
        "description": "Retrieves a specific question from a test suite by its ID",
        "operationId": "getById",
        "parameters": [
          {
            "name": "testSuiteId",
            "in": "path",
            "description": "Test Suite ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Question ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174001"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "200": {
            "description": "Question retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "404": {
            "description": "Question or test suite not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Question"
        ],
        "summary": "Update question",
        "description": "Updates an existing question in a specific test suite",
        "operationId": "update",
        "parameters": [
          {
            "name": "testSuiteId",
            "in": "path",
            "description": "Test Suite ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Question ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174001"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Question updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or metadata validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "404": {
            "description": "Question or test suite not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Question"
        ],
        "summary": "Delete question",
        "description": "Soft deletes a question from a specific test suite",
        "operationId": "deleteById",
        "parameters": [
          {
            "name": "testSuiteId",
            "in": "path",
            "description": "Test Suite ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Question ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174001"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "200": {
            "description": "Question deleted successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "404": {
            "description": "Question or test suite not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          }
        }
      }
    },
    "/shelves/{id}": {
      "get": {
        "tags": [
          "Shelve"
        ],
        "summary": "Get shelve by ID",
        "description": "Retrieves a shelve by its unique identifier",
        "operationId": "getById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Shelve ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "404": {
            "description": "Shelve not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "200": {
            "description": "Shelve retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Shelve"
        ],
        "summary": "Update shelve",
        "description": "Updates an existing shelve by ID",
        "operationId": "update_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Shelve ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShelveReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid request body or ID",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "404": {
            "description": "Shelve not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "200": {
            "description": "Shelve updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Shelve"
        ],
        "summary": "Delete shelve",
        "description": "Soft deletes a shelve by ID",
        "operationId": "deleteById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Shelve ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "responses": {
          "200": {
            "description": "Shelve deleted successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "404": {
            "description": "Shelve not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          }
        }
      }
    },
    "/shelves/{shelveId}/test-suites/{id}": {
      "get": {
        "tags": [
          "Test Suite"
        ],
        "summary": "Get test suite by ID",
        "description": "Retrieves a specific test suite from a shelve by its ID",
        "operationId": "getById_2",
        "parameters": [
          {
            "name": "shelveId",
            "in": "path",
            "description": "Shelve ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Test Suite ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174001"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "200": {
            "description": "Test suite retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "404": {
            "description": "Test suite or shelve not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Test Suite"
        ],
        "summary": "Update test suite",
        "description": "Updates an existing test suite in a specific shelve",
        "operationId": "update_2",
        "parameters": [
          {
            "name": "shelveId",
            "in": "path",
            "description": "Shelve ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Test Suite ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174001"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestSuiteReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid request body",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "200": {
            "description": "Test suite updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "404": {
            "description": "Test suite or shelve not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Test Suite"
        ],
        "summary": "Delete test suite",
        "description": "Soft deletes a test suite and all its questions from a specific shelve",
        "operationId": "deleteById_2",
        "parameters": [
          {
            "name": "shelveId",
            "in": "path",
            "description": "Shelve ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Test Suite ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174001"
          }
        ],
        "responses": {
          "200": {
            "description": "Test suite deleted successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "404": {
            "description": "Test suite or shelve not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          }
        }
      }
    },
    "/shelves/{shelveId}/test-suites/{id}/run": {
      "post": {
        "tags": [
          "Test Suite"
        ],
        "summary": "Run test preset",
        "description": "Returns flashcards matching this suite's linked subjects and selection rules (read-only snapshot for a practice session)",
        "operationId": "runTestPreset",
        "parameters": [
          {
            "name": "shelveId",
            "in": "path",
            "description": "Shelve ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Test Suite ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174001"
          }
        ],
        "responses": {
          "200": {
            "description": "Cards resolved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "404": {
            "description": "Test suite or shelve not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          }
        }
      }
    },
...
    "/shelves/{shelveId}/test-suites": {
      "get": {
        "tags": [
          "Test Suite"
        ],
        "summary": "Get all test suites by shelve",
        "description": "Retrieves a list of all test suites for a specific shelve",
        "operationId": "getAllByDeck",
        "parameters": [
          {
            "name": "shelveId",
            "in": "path",
            "description": "Shelve ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "responses": {
          "200": {
            "description": "Test suites retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "404": {
            "description": "Shelve not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Test Suite"
        ],
        "summary": "Create test suite",
        "description": "Creates a new test suite within a specific shelve",
        "operationId": "create_2",
        "parameters": [
          {
            "name": "shelveId",
            "in": "path",
            "description": "Shelve ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestSuiteReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid request body",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "201": {
            "description": "Test suite created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          },
          "404": {
            "description": "Shelve not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Res"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "QuestionReq": {
        "type": "object",
        "description": "Question creation request",
        "properties": {
          "text": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "MULTIPLE_CHOICE",
              "MULTIPLE_RESPONSE",
              "TRUE_FALSE",
              "FILL_IN_THE_BLANK",
              "MATCHING",
              "ORDERING"
            ]
          },
          "metadata": {
            "type": "string"
          }
        },
        "required": [
          "metadata",
          "text",
          "type"
        ]
      },
      "Res": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "isSuccess": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ShelveReq": {
        "type": "object",
        "description": "Shelve creation request",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "maxLength": 100,
            "minLength": 10
          }
        }
      },
      "TestSuiteReq": {
        "type": "object",
        "description": "Test suite preset: title, optional subject scope, and card selection rules",
        "properties": {
          "title": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "subjectIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "selection": {
            "$ref": "#/components/schemas/TestSuiteSelectionPayload"
          }
        },
        "required": [
          "title"
        ]
      },
      "TestSuiteSelectionPayload": {
        "type": "object",
        "description": "How to draw cards when running a test preset",
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "ALL",
              "DUE_ONLY",
              "RANDOM"
            ],
            "default": "ALL"
          },
          "limit": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "shuffle": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
    }
  }
}